+++ confd/warp/warp.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/cloudogu/ces-confd/confd"
 	"github.com/cloudogu/ces-confd/confd/registry"
-	"github.com/coreos/etcd/client"
 	"github.com/pkg/errors"
+	"go.etcd.io/etcd/client/v2"
 )
 
 // Configuration for warp menu creation