+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Upgrade go-version to v1.23.5 to fix CVEs in v1.17.13 which no longer receives updates (#36)
+  - Replace the `github.com/coreos/etcd/client`-module with `go.etcd.io/etcd/client/v2` to be compatible with new go versions
 
 ## [v0.10.0] - 2024-09-18
 ### Changed